So, it turns out that attempting to write a little language in which
  keyword => ...

or

  keyword : ...

want to be legal isn't ever going to fly in perl because, in the case of
=>, the keyword is auto stringified before Devel::Declare attempts to do
its thing, and 'keyword :' is interpreted as a label, before Devel::Declare
gets a sniff at things.

So... welcome one and all the new mechanism for setting your plan:

setSo, it turns out that attempting to write a little language in which

  keyword => ...

or

  keyword : ...

want to be legal isn't ever going to fly in perl because, in the case of
=>, the keyword is auto stringified before Devel::Declare attempts to do
its thing, and 'keyword :' is interpreted as a label, before Devel::Declare
gets a sniff at things.

So... welcome one and all the new mechanism for setting your plan:

  setup >> 1 {
    ...
  }

You'll note too that that 'setup' doesn't have a name. So Test::Class::Sugar
invents one based on the name of the test class.

  setup  >> 1 {
    ...
  }

You'll note too that that 'setup' doesn't have a name. So Test::Class::Sugar
invents one based on the name of the test class. So,

  testclass Test::Foo {
    setup >> 1 {
      ...
    }
  }

is equivalent to

  testclass Test::Foo {
    setup Test Foo >> 1 {
      ...
    }
  }

Which is, in turn, equivalent to:

  testclass Test::Foo {
    sub setup_Test_Foo :Test(1) {
      ...
    }
  }

And everyone's happy.

Wow. Longest commit message ever!